+++ pkgs/dart_mcp/lib/src/api/tools.dart
@@ -45,7 +45,23 @@ extension type ListToolsResult.fromMap(Map<String, Object?> _value)
   }
 }
 
-/// The server's response to a tool call.
+/// A response to a `tools/call` request, which is either a [CallToolResult]
+/// or an [InputRequiredResult].
+///
+/// Check [isInputRequired] and then cast to one of those.
+///
+/// This type is not intended to be constructed directly and thus has no public
+/// constructor.
+extension type CallToolResponse._fromMap(Map<String, Object?> _value)
+    implements Result {
+  /// A complete result with no input required.
+  static const callToolResult = CallToolResult.new;
+
+  /// A result that requires further input from the client.
+  static const inputRequiredResult = InputRequiredResult.new;
+}
+
+/// The server's completed response to a tool call.
 ///
 /// Any errors that originate from the tool SHOULD be reported inside the result
 /// object, with `isError` set to true, _not_ as an MCP protocol-level error
@@ -56,7 +72,7 @@ extension type ListToolsResult.fromMap(Map<String, Object?> _value)
 /// server does not support tool calls, or any other exceptional conditions,
 /// should be reported as an MCP error response.
 extension type CallToolResult.fromMap(Map<String, Object?> _value)
-    implements Result {
+    implements CallToolResponse {
   factory CallToolResult({
     Meta? meta,
     required List<Content> content,